quant-trading-bot/requirements.txt
pandas
numpy
requests
python-dotenv
alpaca-trade-api
matplotlib
